contentcollector: Fix Node attribute accesses
The DOM spec doesn't say anything about an `attribs` property on DOM
Node or Element objects, and major browsers don't seem to set an
`attribs` property, so `node.attribs` is always falsy. I believe that
`attributes` was intended (that `attribs` was a typo). Use
`getAttribute()` to hopefully avoid future confusion.