contentcollector: Fix Element tag name fetch
The DOM spec doesn't say anything about a `name` property on DOM Node
or Element objects, and major browsers don't seem to set a `name`
property, so `node.name` is always undefined. I believe that `tagName`
was intended (that `name` was a typo). Switch to the `tagName()`
function for consistency with the rest of the file.